Render two classes public for reuse in applications.
