+++ core/src/main/java/com/yahoo/ycsb/Client.java
@@ -28,296 +28,10 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.text.DecimalFormat;
-import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.Map.Entry;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.LockSupport;
-
-/**
- * A thread to periodically show the status of the experiment to reassure you that progress is being made.
- */
-class StatusThread extends Thread {
-  // Counts down each of the clients completing
-  private final CountDownLatch completeLatch;
-
-  // Stores the measurements for the run
-  private final Measurements measurements;
-
-  // Whether or not to track the JVM stats per run
-  private final boolean trackJVMStats;
-
-  // The clients that are running.
-  private final List<ClientThread> clients;
-
-  private final String label;
-  private final boolean standardstatus;
-
-  // The interval for reporting status.
-  private long sleeptimeNs;
-
-  // JVM max/mins
-  private int maxThreads;
-  private int minThreads = Integer.MAX_VALUE;
-  private long maxUsedMem;
-  private long minUsedMem = Long.MAX_VALUE;
-  private double maxLoadAvg;
-  private double minLoadAvg = Double.MAX_VALUE;
-  private long lastGCCount = 0;
-  private long lastGCTime = 0;
-
-  /**
-   * Creates a new StatusThread without JVM stat tracking.
-   *
-   * @param completeLatch         The latch that each client thread will {@link CountDownLatch#countDown()}
-   *                              as they complete.
-   * @param clients               The clients to collect metrics from.
-   * @param label                 The label for the status.
-   * @param standardstatus        If true the status is printed to stdout in addition to stderr.
-   * @param statusIntervalSeconds The number of seconds between status updates.
-   */
-  public StatusThread(CountDownLatch completeLatch, List<ClientThread> clients,
-                      String label, boolean standardstatus, int statusIntervalSeconds) {
-    this(completeLatch, clients, label, standardstatus, statusIntervalSeconds, false);
-  }
-
-  /**
-   * Creates a new StatusThread.
-   *
-   * @param completeLatch         The latch that each client thread will {@link CountDownLatch#countDown()}
-   *                              as they complete.
-   * @param clients               The clients to collect metrics from.
-   * @param label                 The label for the status.
-   * @param standardstatus        If true the status is printed to stdout in addition to stderr.
-   * @param statusIntervalSeconds The number of seconds between status updates.
-   * @param trackJVMStats         Whether or not to track JVM stats.
-   */
-  public StatusThread(CountDownLatch completeLatch, List<ClientThread> clients,
-                      String label, boolean standardstatus, int statusIntervalSeconds,
-                      boolean trackJVMStats) {
-    this.completeLatch = completeLatch;
-    this.clients = clients;
-    this.label = label;
-    this.standardstatus = standardstatus;
-    sleeptimeNs = TimeUnit.SECONDS.toNanos(statusIntervalSeconds);
-    measurements = Measurements.getMeasurements();
-    this.trackJVMStats = trackJVMStats;
-  }
-
-  /**
-   * Run and periodically report status.
-   */
-  @Override
-  public void run() {
-    final long startTimeMs = System.currentTimeMillis();
-    final long startTimeNanos = System.nanoTime();
-    long deadline = startTimeNanos + sleeptimeNs;
-    long startIntervalMs = startTimeMs;
-    long lastTotalOps = 0;
-
-    boolean alldone;
-
-    do {
-      long nowMs = System.currentTimeMillis();
-
-      lastTotalOps = computeStats(startTimeMs, startIntervalMs, nowMs, lastTotalOps);
-
-      if (trackJVMStats) {
-        measureJVM();
-      }
-
-      alldone = waitForClientsUntil(deadline);
-
-      startIntervalMs = nowMs;
-      deadline += sleeptimeNs;
-    }
-    while (!alldone);
-
-    if (trackJVMStats) {
-      measureJVM();
-    }
-    // Print the final stats.
-    computeStats(startTimeMs, startIntervalMs, System.currentTimeMillis(), lastTotalOps);
-  }
-
-  /**
-   * Computes and prints the stats.
-   *
-   * @param startTimeMs     The start time of the test.
-   * @param startIntervalMs The start time of this interval.
-   * @param endIntervalMs   The end time (now) for the interval.
-   * @param lastTotalOps    The last total operations count.
-   * @return The current operation count.
-   */
-  private long computeStats(final long startTimeMs, long startIntervalMs, long endIntervalMs,
-                            long lastTotalOps) {
-    SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss:SSS");
-
-    long totalops = 0;
-    long todoops = 0;
-
-    // Calculate the total number of operations completed.
-    for (ClientThread t : clients) {
-      totalops += t.getOpsDone();
-      todoops += t.getOpsTodo();
-    }
-
-
-    long interval = endIntervalMs - startTimeMs;
-    double throughput = 1000.0 * (((double) totalops) / (double) interval);
-    double curthroughput = 1000.0 * (((double) (totalops - lastTotalOps)) /
-        ((double) (endIntervalMs - startIntervalMs)));
-    long estremaining = (long) Math.ceil(todoops / throughput);
-
-
-    DecimalFormat d = new DecimalFormat("#.##");
-    String labelString = this.label + format.format(new Date());
-
-    StringBuilder msg = new StringBuilder(labelString).append(" ").append(interval / 1000).append(" sec: ");
-    msg.append(totalops).append(" operations; ");
-
-    if (totalops != 0) {
-      msg.append(d.format(curthroughput)).append(" current ops/sec; ");
-    }
-    if (todoops != 0) {
-      msg.append("est completion in ").append(RemainingFormatter.format(estremaining));
-    }
-
-    msg.append(Measurements.getMeasurements().getSummary());
-
-    System.err.println(msg);
-
-    if (standardstatus) {
-      System.out.println(msg);
-    }
-    return totalops;
-  }
-
-  /**
-   * Waits for all of the client to finish or the deadline to expire.
-   *
-   * @param deadline The current deadline.
-   * @return True if all of the clients completed.
-   */
-  private boolean waitForClientsUntil(long deadline) {
-    boolean alldone = false;
-    long now = System.nanoTime();
-
-    while (!alldone && now < deadline) {
-      try {
-        alldone = completeLatch.await(deadline - now, TimeUnit.NANOSECONDS);
-      } catch (InterruptedException ie) {
-        // If we are interrupted the thread is being asked to shutdown.
-        // Return true to indicate that and reset the interrupt state
-        // of the thread.
-        Thread.currentThread().interrupt();
-        alldone = true;
-      }
-      now = System.nanoTime();
-    }
-
-    return alldone;
-  }
-
-  /**
-   * Executes the JVM measurements.
-   */
-  private void measureJVM() {
-    final int threads = Utils.getActiveThreadCount();
-    if (threads < minThreads) {
-      minThreads = threads;
-    }
-    if (threads > maxThreads) {
-      maxThreads = threads;
-    }
-    measurements.measure("THREAD_COUNT", threads);
-
-    // TODO - once measurements allow for other number types, switch to using
-    // the raw bytes. Otherwise we can track in MB to avoid negative values 
-    // when faced with huge heaps.
-    final int usedMem = Utils.getUsedMemoryMegaBytes();
-    if (usedMem < minUsedMem) {
-      minUsedMem = usedMem;
-    }
-    if (usedMem > maxUsedMem) {
-      maxUsedMem = usedMem;
-    }
-    measurements.measure("USED_MEM_MB", usedMem);
-
-    // Some JVMs may not implement this feature so if the value is less than
-    // zero, just ommit it.
-    final double systemLoad = Utils.getSystemLoadAverage();
-    if (systemLoad >= 0) {
-      // TODO - store the double if measurements allows for them
-      measurements.measure("SYS_LOAD_AVG", (int) systemLoad);
-      if (systemLoad > maxLoadAvg) {
-        maxLoadAvg = systemLoad;
-      }
-      if (systemLoad < minLoadAvg) {
-        minLoadAvg = systemLoad;
-      }
-    }
-
-    final long gcs = Utils.getGCTotalCollectionCount();
-    measurements.measure("GCS", (int) (gcs - lastGCCount));
-    final long gcTime = Utils.getGCTotalTime();
-    measurements.measure("GCS_TIME", (int) (gcTime - lastGCTime));
-    lastGCCount = gcs;
-    lastGCTime = gcTime;
-  }
-
-  /**
-   * @return The maximum threads running during the test.
-   */
-  public int getMaxThreads() {
-    return maxThreads;
-  }
-
-  /**
-   * @return The minimum threads running during the test.
-   */
-  public int getMinThreads() {
-    return minThreads;
-  }
-
-  /**
-   * @return The maximum memory used during the test.
-   */
-  public long getMaxUsedMem() {
-    return maxUsedMem;
-  }
-
-  /**
-   * @return The minimum memory used during the test.
-   */
-  public long getMinUsedMem() {
-    return minUsedMem;
-  }
-
-  /**
-   * @return The maximum load average during the test.
-   */
-  public double getMaxLoadAvg() {
-    return maxLoadAvg;
-  }
-
-  /**
-   * @return The minimum load average during the test.
-   */
-  public double getMinLoadAvg() {
-    return minLoadAvg;
-  }
-
-  /**
-   * @return Whether or not the thread is tracking JVM stats.
-   */
-  public boolean trackJVMStats() {
-    return trackJVMStats;
-  }
-}
 
 /**
  * Turn seconds remaining into more useful units.
@@ -353,168 +67,6 @@ final class RemainingFormatter {
       time.append(seconds).append(time.length() == 1 ? " second " : " seconds ");
     }
     return time;
-  }
-}
-
-/**
- * A thread for executing transactions or data inserts to the database.
- */
-class ClientThread implements Runnable {
-  // Counts down each of the clients completing.
-  private final CountDownLatch completeLatch;
-
-  private static boolean spinSleep;
-  private DB db;
-  private boolean dotransactions;
-  private Workload workload;
-  private int opcount;
-  private double targetOpsPerMs;
-
-  private int opsdone;
-  private int threadid;
-  private int threadcount;
-  private Object workloadstate;
-  private Properties props;
-  private long targetOpsTickNs;
-  private final Measurements measurements;
-
-  /**
-   * Constructor.
-   *
-   * @param db                   the DB implementation to use
-   * @param dotransactions       true to do transactions, false to insert data
-   * @param workload             the workload to use
-   * @param props                the properties defining the experiment
-   * @param opcount              the number of operations (transactions or inserts) to do
-   * @param targetperthreadperms target number of operations per thread per ms
-   * @param completeLatch        The latch tracking the completion of all clients.
-   */
-  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, int opcount,
-                      double targetperthreadperms, CountDownLatch completeLatch) {
-    this.db = db;
-    this.dotransactions = dotransactions;
-    this.workload = workload;
-    this.opcount = opcount;
-    opsdone = 0;
-    if (targetperthreadperms > 0) {
-      targetOpsPerMs = targetperthreadperms;
-      targetOpsTickNs = (long) (1000000 / targetOpsPerMs);
-    }
-    this.props = props;
-    measurements = Measurements.getMeasurements();
-    spinSleep = Boolean.valueOf(this.props.getProperty("spin.sleep", "false"));
-    this.completeLatch = completeLatch;
-  }
-
-  public void setThreadId(final int threadId) {
-    threadid = threadId;
-  }
-  
-  public void setThreadCount(final int threadCount) {
-    threadcount = threadCount;
-  }
-  
-  public int getOpsDone() {
-    return opsdone;
-  }
-
-  @Override
-  public void run() {
-    try {
-      db.init();
-    } catch (DBException e) {
-      e.printStackTrace();
-      e.printStackTrace(System.out);
-      return;
-    }
-
-    try {
-      workloadstate = workload.initThread(props, threadid, threadcount);
-    } catch (WorkloadException e) {
-      e.printStackTrace();
-      e.printStackTrace(System.out);
-      return;
-    }
-
-    //NOTE: Switching to using nanoTime and parkNanos for time management here such that the measurements
-    // and the client thread have the same view on time.
-
-    //spread the thread operations out so they don't all hit the DB at the same time
-    // GH issue 4 - throws exception if _target>1 because random.nextInt argument must be >0
-    // and the sleep() doesn't make sense for granularities < 1 ms anyway
-    if ((targetOpsPerMs > 0) && (targetOpsPerMs <= 1.0)) {
-      long randomMinorDelay = ThreadLocalRandom.current().nextInt((int) targetOpsTickNs);
-      sleepUntil(System.nanoTime() + randomMinorDelay);
-    }
-    try {
-      if (dotransactions) {
-        long startTimeNanos = System.nanoTime();
-
-        while (((opcount == 0) || (opsdone < opcount)) && !workload.isStopRequested()) {
-
-          if (!workload.doTransaction(db, workloadstate)) {
-            break;
-          }
-
-          opsdone++;
-
-          throttleNanos(startTimeNanos);
-        }
-      } else {
-        long startTimeNanos = System.nanoTime();
-
-        while (((opcount == 0) || (opsdone < opcount)) && !workload.isStopRequested()) {
-
-          if (!workload.doInsert(db, workloadstate)) {
-            break;
-          }
-
-          opsdone++;
-
-          throttleNanos(startTimeNanos);
-        }
-      }
-    } catch (Exception e) {
-      e.printStackTrace();
-      e.printStackTrace(System.out);
-      System.exit(0);
-    }
-
-    try {
-      measurements.setIntendedStartTimeNs(0);
-      db.cleanup();
-    } catch (DBException e) {
-      e.printStackTrace();
-      e.printStackTrace(System.out);
-    } finally {
-      completeLatch.countDown();
-    }
-  }
-
-  private static void sleepUntil(long deadline) {
-    while (System.nanoTime() < deadline) {
-      if (!spinSleep) {
-        LockSupport.parkNanos(deadline - System.nanoTime());
-      }
-    }
-  }
-
-  private void throttleNanos(long startTimeNanos) {
-    //throttle the operations
-    if (targetOpsPerMs > 0) {
-      // delay until next tick
-      long deadline = startTimeNanos + opsdone * targetOpsTickNs;
-      sleepUntil(deadline);
-      measurements.setIntendedStartTimeNs(deadline);
-    }
-  }
-
-  /**
-   * The total amount of work this thread is still expected to do.
-   */
-  int getOpsTodo() {
-    int todo = opcount - opsdone;
-    return todo < 0 ? 0 : todo;
   }
 }
